+++ server/src/schedulers/scheduler_chunks.rs
@@ -19,9 +19,9 @@ pub struct GlobalConfiguration {
     #[primary_key]
     #[auto_inc]
     #[wrap]    
-    id: u64,
-    pub scheduled_id: u64,
-    pub current_update: u64,
+    id: u32,
+    pub scheduled_id: u32,
+    pub current_update: u32,
 }
 
 #[dsl(plural_name = chunk_check_timers)]
@@ -30,8 +30,8 @@ pub struct ChunkCheckTimer {
     #[primary_key]
     #[auto_inc]
     #[wrap]    
-    id: u64,
-    pub scheduled_id: u64,
+    id: u32,
+    pub scheduled_id: u32,
     scheduled_at: spacetimedb::ScheduleAt,
     current_update: u8,
 }
